Encapsulate ListerDetails Component from AgentFloatingWidget
