+++ src/components/property-detail/AgentFloatingWidget.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { Star } from '@styled-icons/boxicons-solid';
+import ListerDetails from "./ListerDetails";
 
 const Container = styled.div`
   width: 300px;
@@ -18,79 +18,15 @@ const BrandingBar = styled.div`
   background-color: ${props => props.backgroundColor};
 `;
 
-const AgentPhotoUrl = styled.img`
-  border: 1px;
-  border-style: solid;
-  border-color: #D2D5DA;
-  border-radius: 50%;
-  height: 60px;
-  width: 60px;
+const Divider = styled.div`
+  width: 300px;
+  height: 1px;
+  background-color: #D2D5DA;
 `;
-
-const AgentDetailsContainer = styled.div`
-  display: flex;
-  justify-content: flex-start;
-`;
-
-const AgentPhotoUrlContainer = styled.div`
-  margin: 18px 16px;
-`;
-
-const AgentDetails = styled.div`
-  margin: 16px 0;
-`;
-
-const AgentName = styled.div`
-  font-family: "PangeaRegular";
-  font-size: 16px;
-  font-weight: 400;
-  line-height: 24px;
-  color: #2B6ED2;
-`;
-
-const AgentNameLink = styled.a`
-  text-decoration: none;
-  &:visited {
-    color: #2B6ED2
-  };
-  &:hover {
-    text-decoration: underline;
-  }
-`;
-
-const ListerPhoneNumber = styled.div`
-  font-family: "PangeaRegular";
-  font-size: 12px;
-  font-weight: 300;
-  line-height: 16px;
-  color: #333F48;
-`;
-
-const ListerRatingsReviewsContainer = styled.div`
-  font-family: "PangeaRegular";
-  font-size: 14px;
-  font-weight: 300;
-  line-height: 20px;
-  color: #697684;
-`;
-
-const ListerRatingsReviews = (props) => {
-  if (props.listerRatingsReviews.totalReviews === 0) {
-    return <ListerRatingsReviewsContainer>(No reviews yet)</ListerRatingsReviewsContainer>
-  } else {
-    return(
-      <ListerRatingsReviewsContainer>
-          <Star size="14" color="#FFB200" /> {props.listerRatingsReviews.avgRating} ({props.listerRatingsReviews.totalReviews} reviews)
-      </ListerRatingsReviewsContainer>
-    );
-  }
-};
 
 const AgentFloatingWidget = (props) => {
   const logoUrl = props.listingCompany.media.logo.templatedUrl.replace('{size}', '170x32');
-  const agentPhotoUrl = props.listers[0].photo.templatedUrl.replace('{size}', '120x120');
-  // const listerRatingsReviews = props.listers[0].listerRatingsReviews;
-  const listerRatingsReviews = { avgRating: 4.5, totalReviews: 32 };
+
   return(
     <Container>
       <BrandingBar backgroundColor={props.listingCompany.branding.primaryColour}>
@@ -98,18 +34,12 @@ const AgentFloatingWidget = (props) => {
           <img src={logoUrl} alt="" />
         </a>
       </BrandingBar>
-      <AgentDetailsContainer>
-        <AgentPhotoUrlContainer>
-          <AgentPhotoUrl src={agentPhotoUrl} alt="" />
-        </AgentPhotoUrlContainer>
-        <AgentDetails>
-           <AgentName>
-             <AgentNameLink target="_blank" href={props.listingCompany._links.canonical.href}>{props.listers[0].name}</AgentNameLink>
-            </AgentName>
-           <ListerRatingsReviews listerRatingsReviews={listerRatingsReviews} />
-           <ListerPhoneNumber>{props.listers[0].phoneNumber.display}</ListerPhoneNumber>
-        </AgentDetails>
-      </AgentDetailsContainer>
+      {
+        props.listers.map((item) => {
+          return <ListerDetails lister={item} />
+        })
+      }
+      <Divider />
     </Container>
   );
 }
